refactor: extract router setup and fix compiler and test warnings
