+++ backend/docs/docs.go
@@ -1,0 +1,3 @@
+package docs
+
+// Docs package placeholders for Swagger
