+++ backend/monitoring/tracing.go
@@ -4,11 +4,10 @@ import (
 	"context"
 	"log"
 
-	"go.opentelemetry.io/exporter/otlp/otlptrace"
-	"go.opentelemetry.io/exporter/otlp/otlptrace/otlptracehttp"
-	"go.opentelemetry.io/sdk/resource"
-	sdktrace "go.opentelemetry.io/sdk/trace"
-	semconv "go.opentelemetry.io/semconv/v1.17.0"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
+	"go.opentelemetry.io/otel/sdk/resource"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 )
 
 func InitializeTracing(ctx context.Context, serviceName, collectorURL string) (*sdktrace.TracerProvider, error) {
