+++ backend/tests/integration/test_helper.go
@@ -12,15 +12,13 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/joho/godotenv"
 	"github.com/stellar/go/keypair"
 	"github.com/stretchr/testify/require"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 
-	"github.com/yourusername/kor-assetforge/config"
 	"github.com/yourusername/kor-assetforge/models"
-	"github.com/yourusername/kor-assetforge/main"
+	"github.com/yourusername/kor-assetforge/router"
 )
 
 // TestSetup provides test environment setup and utilities
@@ -55,17 +53,15 @@ func NewTestSetup(t *testing.T) *TestSetup {
 	err = db.AutoMigrate(
 		&models.Asset{},
 		&models.User{},
-		&models.KYC{},
+		&models.KYCRecord{},
 		&models.Listing{},
 		&models.Transaction{},
-		&models.GovernanceProposal{},
-		&models.EmergencyControl{},
 	)
 	require.NoError(t, err)
 
 	// Setup Gin in test mode
 	gin.SetMode(gin.TestMode)
-	server := main.SetupRouter(db)
+	server := router.SetupRouter(db)
 
 	// Create admin user
 	adminUser := createAdminUser(t, db)
@@ -75,7 +71,7 @@ func NewTestSetup(t *testing.T) *TestSetup {
 		DB:     db,
 		Server: server,
 		TestUsers: []TestUser{},
-		AdminToken: generateTestToken(adminUser.PublicKey),
+		AdminToken: generateTestToken(adminUser.StellarAddress),
 		CleanupFunc: func() {
 			// Cleanup database
 			sqlDB, _ := db.DB()
@@ -104,18 +100,20 @@ func (ts *TestSetup) CreateTestUser(t *testing.T) *TestUser {
 
 	// Create user in database
 	user := &models.User{
-		PublicKey: pair.Address(),
-		Role:      "user",
-		Status:    "active",
+		StellarAddress: pair.Address(),
+		Role:           models.RoleUser,
+		Email:          fmt.Sprintf("user_%s@example.com", pair.Address()[:8]),
+		Username:       fmt.Sprintf("user_%s", pair.Address()[:8]),
 	}
 
 	err = ts.DB.Create(user).Error
 	require.NoError(t, err)
 
 	// Create KYC record
-	kyc := &models.KYC{
-		UserID: user.ID,
-		Status: "pending",
+	kyc := &models.KYCRecord{
+		UserID:   user.ID,
+		Status:   models.KYCStatusPending,
+		FullName: "Test User",
 	}
 
 	err = ts.DB.Create(kyc).Error
@@ -157,12 +155,12 @@ func (ts *TestSetup) MakeRequest(method, path string, body interface{}, token st
 }
 
 // ApproveKYC approves KYC for a user (admin action)
-func (ts *TestSetup) ApproveKYC(t *testing.T, publicKey string) {
+func (ts *TestSetup) ApproveKYC(t *testing.T, stellarAddress string) {
 	var user models.User
-	err := ts.DB.Where("public_key = ?", publicKey).First(&user).Error
+	err := ts.DB.Where("stellar_address = ?", stellarAddress).First(&user).Error
 	require.NoError(t, err)
 
-	err = ts.DB.Model(&models.KYC{}).Where("user_id = ?", user.ID).Update("status", "approved").Error
+	err = ts.DB.Model(&models.KYCRecord{}).Where("user_id = ?", user.ID).Update("status", models.KYCStatusApproved).Error
 	require.NoError(t, err)
 }
 
@@ -172,19 +170,19 @@ func (ts *TestSetup) seedTestData(t *testing.T) {
 	assets := []models.Asset{
 		{
 			Name:        "Test Property 1",
-			Code:        "TEST1",
+			Symbol:      "TEST1",
 			Description: "Test property for integration testing",
 			TotalSupply: 1000000,
-			Decimals:    7,
-			Status:      "active",
+			Fractions:   7,
+			ContractID:  "C1",
 		},
 		{
 			Name:        "Test Property 2",
-			Code:        "TEST2",
+			Symbol:      "TEST2",
 			Description: "Another test property",
 			TotalSupply: 500000,
-			Decimals:    7,
-			Status:      "active",
+			Fractions:   7,
+			ContractID:  "C2",
 		},
 	}
 
@@ -200,9 +198,10 @@ func createAdminUser(t *testing.T, db *gorm.DB) *models.User {
 	require.NoError(t, err)
 
 	admin := &models.User{
-		PublicKey: pair.Address(),
-		Role:      "admin",
-		Status:    "active",
+		StellarAddress: pair.Address(),
+		Role:           models.RoleAdmin,
+		Email:          fmt.Sprintf("admin_%s@example.com", pair.Address()[:8]),
+		Username:       fmt.Sprintf("admin_%s", pair.Address()[:8]),
 	}
 
 	err = db.Create(admin).Error
