+++ backend/validator/validator_test.go
@@ -19,7 +19,7 @@ func TestValidateTokenizeAssetRequest(t *testing.T) {
     }
 
     req := TokenizeAssetRequest{
-        IssuerAccount: "GD6WU5I6OIPRZ4A5I3G6JQ4RG5K27SQ26WPQ5W3MXV6QABBT3C7FIEIF",
+        IssuerAccount: "GABXYMNLGGTWAV7EQYHVWLQJ7MSAKBW3OX4J5B3UALQTVOUGX3HXBMEM",
         Name:          "Real Asset",
         Symbol:        "RWA1",
         AssetType:     "real_estate",
@@ -37,7 +37,7 @@ func TestValidateTokenizeAssetRequestRejectsHtml(t *testing.T) {
     }
 
     req := TokenizeAssetRequest{
-        IssuerAccount: "GD6WU5I6OIPRZ4A5I3G6JQ4RG5K27SQ26WPQ5W3MXV6QABBT3C7FIEIF",
+        IssuerAccount: "GABXYMNLGGTWAV7EQYHVWLQJ7MSAKBW3OX4J5B3UALQTVOUGX3HXBMEM",
         Name:          "Real <b>Asset</b>",
         Symbol:        "RWA1",
         AssetType:     "real_estate",
